refactor(preflight): integrate HealthAggregator and MedallionConfigValidator
Integrate HealthAggregator and MedallionConfigValidator classes directly into
PreflightService as private helper classes (_HealthAggregator and
_MedallionConfigValidator). This eliminates redundant public API surface since
these classes were only used internally by PreflightService.

Changes:
- Move HealthAggregator logic to _HealthAggregator within preflight_service.py
- Move MedallionConfigValidator logic to _MedallionConfigValidator
- Remove health_aggregator.py and medallion_validator.py files
- Remove HealthAggregator from __init__.py public exports
- Update tests to import from new location

Benefits:
- PreflightService is now self-contained
- Reduced public API surface in application/core
- All preflight validation logic in one module
- Preserved testability via private class imports
